refactor(widget): rename mis-named utils so the name states the concern
utils/common.ts is solely a scoped-logger factory → logger.ts.
utils/format.ts is ~85% color math (hexToRgb/getContrastingColor/addOpacity/
darkenColor) → color.ts.

Names now describe content (root discipline #11). Importers rewired.
